Setup: Moved source code files to package directory and modified setup.py
