+++ sdc_api_py/_types/Guild.py
