+++ setup.py
@@ -12,7 +12,7 @@ setuptools.setup(
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://github.com/TheKing-OfTime/sdc-api.py",
-    packages=setuptools.find_packages(),
+    packages=["sdc_api_py"],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
